miniRT

+++ sphere.h
@@ -10,7 +10,7 @@ typedef struct s_sphere {
     double rad;
 } t_sphere;
 
-double hit_sphere(t_sphere* s, t_ray* r, t_record* rec);
+int hit_sphere(t_sphere* s, t_ray* r, t_record* rec);
 t_sphere create_sphere(t_point c, double r);
 
 #endif
